requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
distro==1.9.0
filelock==3.25.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.4
openai==2.31.0
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
rich==15.0.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
tabulate==0.10.0
together==2.7.0
tqdm==4.67.3
types-PyYAML==6.0.12.20260408
types-requests==2.33.0.20260408
types-tabulate==0.10.0.20260408
types-tqdm==4.67.3.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
